Wrap play page client in Suspense
